Start adding qwerty mode to my layout
Implement a basic qwerty mode by using layers and combos.  This also
implements "home row" modifiers, although instead of the normal home
row, which would be pairs of keys, we implement it just with the bottom
row
     Z   X   C    V        M    ,    .   /
    GUI ALT CTRL SHIFT   SHIFT CTRL ALT GUI

This has a number mode for the top for, with the bottom row shifted,
which might be easier than using the shift keys, which have to be
pressed before the num key.