d3zd3z/proto2: Enable split keyboard mode
Enable split keyboard mode so both halves work.  The sides are
distinguished by a pull up resistor that is populated or not (the GPIO's
seem to be configured with a weak pulldown).

Swap the columns and rows, since the split mode assumes that it is the
rows that are reduced.  This is easily fixed by transposing the LAYOUT
macro.