proto3: Change the modifiers to use a pair of keys
Before this change, the modifiers were causing lots of rollover issues,
which usually resulted in chaotic modified keypress, which often did
unexpected things.

Instead of using the bottom row as the modifiers, instead use the middle
and bottom together for this.  This makes the modifiers a little slower
to press, but allows me to type at full speed using the regular qwerty
keys.