proto3: Move the modifier keys up
Change the modifier keys to use the upper and middle rows.  This allows
the shifted symbol keys below that to work as normal.  A future
consideration might be to make this work with the shifted keys, but
apparently MT and the modified keys conflict with each other.  This
works for now, even though it is a little bit more of a stretch.